ai/neurogenomic_prototype_kit/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
pyyaml==6.0.2
pypdf==4.3.1
python-docx==1.1.2
chromadb==0.5.5
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
transformers==4.43.3
accelerate==0.33.0
peft==0.11.1
bitsandbytes==0.43.1
torch==2.3.1
streamlit==1.37.1